feat: permission check for artwork/[id] routes

+++ app/api/artworks/[id]/route.ts
@@ -1,8 +1,9 @@
 // pages/api/artworks/[id].ts
 import { NextResponse } from 'next/server'
-import { deleteArtworkById, getArtworkById } from '@/lib/queries/artwork'
 import { prisma } from '@/lib/prisma'
-import { Artwork } from '@prisma/client'
+import { Artwork, ArtworkStatus } from '@prisma/client'
+import { getServerSession } from 'next-auth/next'
+import { authOptions } from '@/lib/auth'
 
 // TASK FOR DEV: Auth middleware is needed for PATCH/DELETE.
 export async function GET(
@@ -10,8 +11,18 @@ export async function GET(
   { params }: { params: Promise<{ id: string }> }
 ) {
   const id = (await params).id
+
   try {
-    // const artworks: Artwork[] = await getArtworkById(id)
+    // Check logged in
+    const session = await getServerSession(authOptions)
+
+    if (!session?.user?.id) {
+      return NextResponse.json(
+        { error: 'You must be logged in to get artwork.' },
+        { status: 401 }
+      )
+    }
+
     const artwork: Artwork | null = await prisma.artwork.findUnique({
       where: {
         id: id,
@@ -38,17 +49,51 @@ export async function PATCH(
   req: Request,
   { params }: { params: Promise<{ id: string }> }
 ) {
-  // TASK FOR DEV:
-  // 1. Get user ID from session (MUST be logged in).
-  // 2. Find the artwork.
-  // 3. If artwork.user_id !== userId, return 403 Forbidden.
-  // 4. If artwork.status !== 'PENDING', return 403 (can't edit after approval).
-  // 5. Update the artwork with data from req.body.
-
   const id = (await params).id
-  const body: Partial<Artwork> = await req.json()
   try {
-    // const artworks: Artwork[] = await deleteArtworkById(id)
+    // Check logged in
+    const session = await getServerSession(authOptions)
+
+    if (!session?.user?.id) {
+      return NextResponse.json(
+        { error: 'You must be logged in to update artwork.' },
+        { status: 401 }
+      )
+    }
+
+    // Get Artwork
+    const artwork: Artwork | null = await prisma.artwork.findUnique({
+      where: {
+        id: id,
+      },
+    })
+
+    if (!artwork) {
+      return NextResponse.json(
+        { error: `Artwork with id ${id} not found` },
+        { status: 404 }
+      )
+    }
+
+    // Check if user is allowed to patch artwork
+    if (artwork.user_id !== session.user.id) {
+      return NextResponse.json(
+        { error: `User forbidden from patching artwork with id ${id}` },
+        { status: 403 }
+      )
+    }
+
+    // Check if artwork is pending
+    if (artwork.status === ArtworkStatus.PENDING) {
+      return NextResponse.json(
+        { error: `User cannot edit artwork after approval` },
+        { status: 403 }
+      )
+    }
+
+    // Try to patch artwork
+    const body: Partial<Artwork> = await req.json()
+
     const updatedArtwork: Artwork = await prisma.artwork.update({
       where: {
         id: id,
@@ -69,14 +114,39 @@ export async function DELETE(
   req: Request,
   { params }: { params: Promise<{ id: string }> }
 ) {
-  // TASK FOR DEV:
-  // 1. Get user ID from session (MUST be logged in).
-  // 2. Find the artwork.
-  // 3. If artwork.user_id !== userId, return 403 Forbidden.
-  // 4. Delete the artwork.
   const id = (await params).id
   try {
-    // const artworks: Artwork[] = await deleteArtworkById(id)
+    // Check logged in
+    const session = await getServerSession(authOptions)
+
+    if (!session?.user?.id) {
+      return NextResponse.json(
+        { error: 'You must be logged in to update artwork.' },
+        { status: 401 }
+      )
+    }
+    const artwork: Artwork | null = await prisma.artwork.findUnique({
+      where: {
+        id: id,
+      },
+    })
+
+    if (!artwork) {
+      return NextResponse.json(
+        { error: `Artwork with id ${id} not found` },
+        { status: 404 }
+      )
+    }
+
+    // Check if user is allowed to delete artwork
+    if (artwork.user_id !== session.user.id) {
+      return NextResponse.json(
+        { error: `User forbidden from deleting artwork with id ${id}` },
+        { status: 403 }
+      )
+    }
+
+    // Try to delete artwork
     const deletedArtwork: Artwork = await prisma.artwork.delete({
       where: {
         id: id,
